requirements.txt
langchain==0.2.16
langchain-core==0.2.43
langchain-community==0.2.16
langchain-huggingface==0.0.3

langchain-openai
langchain-pinecone

pinecone-client
openai
tiktoken
python-dotenv
pypdf
sentence-transformers
torch